you can add percussion notes now. And they play from the beginning.
Doesn't work when highlighted a note yet. It will fuck up. But you can
add a percussion note by just doing addNote(note, true); no percussion
note by addNote(note). Percussion note can be any note but for now it
will default to the midi of the synth drum for middle c and will be
displayed in the middle of the bottom clef